# Merkle performance improvements