--- a/arbitrator/bench/src/lib.rs
+++ b/arbitrator/bench/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod parse_input;
+pub mod prepare;