--- a/arbitrator/prover/src/flat_merkle.rs
+++ b/arbitrator/prover/src/flat_merkle.rs
@@ -1,7 +1,7 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-use crate::utils::Bytes32;
+use arbutil::Bytes32;
 use digest::Digest;
 // use rayon::prelude::*;
 use sha3::Keccak256;