--- a/arbitrator/prover/src/lib.rs
+++ b/arbitrator/prover/src/lib.rs
@@ -4,6 +4,7 @@
 #![allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
 
 pub mod binary;
+pub mod flat_merkle;
 mod host;
 #[cfg(feature = "native")]
 mod kzg;
@@ -116,7 +117,7 @@ unsafe fn arbitrator_load_machine_impl(
 pub unsafe extern "C" fn arbitrator_load_wavm_binary(binary_path: *const c_char) -> *mut Machine {
     let binary_path = cstr_to_string(binary_path);
     let binary_path = Path::new(&binary_path);
-    match Machine::new_from_wavm(binary_path) {
+    match Machine::new_from_wavm(binary_path, false) {
         Ok(mach) => Box::into_raw(Box::new(mach)),
         Err(err) => {
             eprintln!("Error loading binary: {err}");