--- a/arbitrator/prover/src/lib.rs
+++ b/arbitrator/prover/src/lib.rs
@@ -101,7 +101,7 @@ unsafe fn arbitrator_load_machine_impl(
         &libraries,
         binary_path,
         true,
-        true,
+        false,
         false,
         debug_chain,
         debug_chain,
@@ -117,7 +117,7 @@ unsafe fn arbitrator_load_machine_impl(
 pub unsafe extern "C" fn arbitrator_load_wavm_binary(binary_path: *const c_char) -> *mut Machine {
     let binary_path = cstr_to_string(binary_path);
     let binary_path = Path::new(&binary_path);
-    match Machine::new_from_wavm(binary_path, true) {
+    match Machine::new_from_wavm(binary_path, false) {
         Ok(mach) => Box::into_raw(Box::new(mach)),
         Err(err) => {
             eprintln!("Error loading binary: {err}");