--- a/arbitrator/prover/src/machine.rs
+++ b/arbitrator/prover/src/machine.rs
@@ -1245,11 +1245,11 @@ impl Machine {
         let soft_float = std::fs::read("../../target/machines/latest/soft-float.wasm")?;
         let soft_float = parse(&soft_float, Path::new("soft-float"))?;
 
-        let mut machine = Self::from_binaries(
+        let mut machine: Machine = Self::from_binaries(
             &[soft_float, wasi_stub, user_test],
             bin,
             false,
-            false,
+            true,
             false,
             compile.debug.debug_funcs,
             true,