This change does two things:

1. It introduces a `Merkle::extend` method on the merkle tree implementation which extends the vectors which hold the leaves and their parents and then calls set on all of the hashes being appended to layer[0].
2. It changes the `Memory::resize` method to use `Merkle::extend` instead of throwing away the merkle tree and creating a new one.

This results in the `always_merkleize` strategy to be much faster than it would be without these changes. Before this change, the `benchbin` binary ran with metrics like these:

```
Running benchmark with always merkleize feature on
avg hash time 49.506µs, avg step time 49.583µs, step size 1, num_iters 100, total time 1.587369292s
avg hash time 131.402µs, avg step time 67.44µs, step size 1024, num_iters 100, total time 1.56975325s
avg hash time 529.716µs, avg step time 65.128135ms, step size 32768, num_iters 100, total time 8.108754125s
avg hash time 486.609µs, avg step time 372.639626ms, step size 1048576, num_iters 100, total time 38.849238375s
```

After this change, the data is more like this:
```
Running benchmark with always merkleize feature on
avg hash time     55.777µs, avg step time     57.294µs, step size        1, num_iters 100, total time  11.315333ms
avg hash time    126.267µs, avg step time     71.536µs, step size     1024, num_iters 100, total time  19.788125ms
avg hash time    497.955µs, avg step time   3.839355ms, step size    32768, num_iters 100, total time  433.74075ms
avg hash time    461.622µs, avg step time  50.498257ms, step size  1048576, num_iters 100, total time  5.09599725s
avg hash time    826.465µs, avg step time 676.037947ms, step size 16777216, num_iters 100, total time 67.686471417s
```

NOTE: With the optimization, even step-size 16,777,216 (2^24) is able to run 100 iterations in just a little bit over a minute.

references https://linear.app/offchain-labs/issue/NIT-2411/arbitrator-optimizations