@eljobe 2024-05-02
I'm closing this PR. The implementation is only fast because it has a bug. The root hash will quite often be incorrect after the merkle tree is extended in `Memory::resize`.
I'll try again after lunch.